services/product_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
certifi==2025.4.26
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.68.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.19.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==1.8.2
pydantic_core==2.33.2
Pygments==2.19.1
pytest==6.2.5
python-dotenv==0.19.0
python-multipart==0.0.5
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.15.0
python-jose==3.3.0
aiofiles==0.7.0
Pillow==9.0.0
requests==2.26.0
